apply fixes to mongodb node
