+++ packages/nodes-base/nodes/MongoDb/MongoDb.node.ts
@@ -3,6 +3,7 @@ import type {
 	FindOneAndUpdateOptions,
 	UpdateOptions,
 	Sort,
+	MongoClient,
 } from 'mongodb';
 import { ObjectId } from 'mongodb';
 import { NodeConnectionTypes, NodeOperationError, UserError } from 'n8n-workflow';
@@ -15,7 +16,6 @@ import type {
 	INodeExecutionData,
 	INodeType,
 	INodeTypeDescription,
-	JsonObject,
 	IPairedItemData,
 } from 'n8n-workflow';
 
@@ -25,6 +25,7 @@ import {
 	parseAndResolveQueryParameters,
 	prepareFields,
 	prepareItems,
+	sanitizeMongoUriInMessage,
 	serializeMongoItems,
 	stringifyObjectIDs,
 	validateAndResolveMongoCredentials,
@@ -64,10 +65,10 @@ export class MongoDb implements INodeType {
 				credential: ICredentialsDecrypted,
 			): Promise<INodeCredentialTestResult> {
 				const credentials = credential.data as IDataObject;
+				let connectionString = '';
 
 				try {
 					const database = ((credentials.database as string) || '').trim();
-					let connectionString = '';
 
 					if (credentials.configurationType === 'connectionString') {
 						connectionString = ((credentials.connectionString as string) || '').trim();
@@ -90,9 +91,10 @@ export class MongoDb implements INodeType {
 					}
 					await client.close();
 				} catch (error) {
+					const errorMessage = error instanceof Error ? error.message : String(error);
 					return {
 						status: 'Error',
-						message: (error as Error).message,
+						message: sanitizeMongoUriInMessage(errorMessage, connectionString),
 					};
 				}
 				return {
@@ -108,7 +110,17 @@ export class MongoDb implements INodeType {
 		const node = this.getNode();
 		const { database, connectionString } = validateAndResolveMongoCredentials(node, credentials);
 		const nodeVersion = node.typeVersion;
-		const client = await connectMongoClient(connectionString, nodeVersion, credentials);
+		const sanitizeErrorMessage = (error: unknown) =>
+			sanitizeMongoUriInMessage(
+				error instanceof Error ? error.message : String(error),
+				connectionString,
+			);
+		let client: MongoClient;
+		try {
+			client = await connectMongoClient(connectionString, nodeVersion, credentials);
+		} catch (error) {
+			throw new NodeOperationError(node, sanitizeErrorMessage(error));
+		}
 		let returnData: INodeExecutionData[] = [];
 
 		try {
@@ -150,7 +162,7 @@ export class MongoDb implements INodeType {
 					} catch (error) {
 						if (this.continueOnFail()) {
 							returnData.push({
-								json: { error: (error as JsonObject).message },
+								json: { error: sanitizeErrorMessage(error) },
 								pairedItem: fallbackPairedItems ?? [{ item: i }],
 							});
 							continue;
@@ -180,7 +192,7 @@ export class MongoDb implements INodeType {
 					} catch (error) {
 						if (this.continueOnFail()) {
 							returnData.push({
-								json: { error: (error as JsonObject).message },
+								json: { error: sanitizeErrorMessage(error) },
 								pairedItem: fallbackPairedItems ?? [{ item: i }],
 							});
 							continue;
@@ -241,7 +253,7 @@ export class MongoDb implements INodeType {
 					} catch (error) {
 						if (this.continueOnFail()) {
 							returnData.push({
-								json: { error: (error as JsonObject).message },
+								json: { error: sanitizeErrorMessage(error) },
 								pairedItem: fallbackPairedItems ?? [{ item: i }],
 							});
 							continue;
@@ -301,7 +313,7 @@ export class MongoDb implements INodeType {
 						} catch (error) {
 							if (this.continueOnFail()) {
 								returnData.push({
-									json: { error: (error as JsonObject).message },
+									json: { error: sanitizeErrorMessage(error) },
 									pairedItem: { item: i },
 								});
 								continue;
@@ -348,7 +360,7 @@ export class MongoDb implements INodeType {
 								.findOneAndReplace(filter, item, updateOptions as FindOneAndReplaceOptions);
 						} catch (error) {
 							if (this.continueOnFail()) {
-								item.json = { error: (error as JsonObject).message };
+								item.json = { error: sanitizeErrorMessage(error) };
 								continue;
 							}
 							throw error;
@@ -413,7 +425,7 @@ export class MongoDb implements INodeType {
 						} catch (error) {
 							if (this.continueOnFail()) {
 								returnData.push({
-									json: { error: (error as JsonObject).message },
+									json: { error: sanitizeErrorMessage(error) },
 									pairedItem: { item: i },
 								});
 								continue;
@@ -461,7 +473,7 @@ export class MongoDb implements INodeType {
 								.findOneAndUpdate(filter, { $set: item }, updateOptions as FindOneAndUpdateOptions);
 						} catch (error) {
 							if (this.continueOnFail()) {
-								item.json = { error: (error as JsonObject).message };
+								item.json = { error: sanitizeErrorMessage(error) };
 								continue;
 							}
 							throw error;
@@ -510,7 +522,7 @@ export class MongoDb implements INodeType {
 						} catch (error) {
 							if (this.continueOnFail()) {
 								returnData.push({
-									json: { error: (error as JsonObject).message },
+									json: { error: sanitizeErrorMessage(error) },
 									pairedItem: { item: i },
 								});
 							} else {
@@ -537,7 +549,7 @@ export class MongoDb implements INodeType {
 							if (this.continueOnFail()) {
 								for (const g of groupItems) {
 									returnData.push({
-										json: { error: (error as JsonObject).message },
+										json: { error: sanitizeErrorMessage(error) },
 										pairedItem: { item: g.originalIndex },
 									});
 								}
@@ -588,7 +600,7 @@ export class MongoDb implements INodeType {
 						}
 					} catch (error) {
 						if (this.continueOnFail()) {
-							responseData = [{ error: (error as JsonObject).message }];
+							responseData = [{ error: sanitizeErrorMessage(error) }];
 						} else {
 							throw error;
 						}
@@ -652,7 +664,7 @@ export class MongoDb implements INodeType {
 						} catch (error) {
 							if (this.continueOnFail()) {
 								returnData.push({
-									json: { error: (error as JsonObject).message },
+									json: { error: sanitizeErrorMessage(error) },
 									pairedItem: { item: i },
 								});
 								continue;
@@ -700,7 +712,7 @@ export class MongoDb implements INodeType {
 								.updateOne(filter, { $set: item }, updateOptions as UpdateOptions);
 						} catch (error) {
 							if (this.continueOnFail()) {
-								item.json = { error: (error as JsonObject).message };
+								item.json = { error: sanitizeErrorMessage(error) };
 								continue;
 							}
 							throw error;
@@ -736,7 +748,7 @@ export class MongoDb implements INodeType {
 					} catch (error) {
 						if (this.continueOnFail()) {
 							returnData.push({
-								json: { error: (error as JsonObject).message },
+								json: { error: sanitizeErrorMessage(error) },
 								pairedItem: fallbackPairedItems ?? [{ item: i }],
 							});
 							continue;
@@ -762,7 +774,7 @@ export class MongoDb implements INodeType {
 					} catch (error) {
 						if (this.continueOnFail()) {
 							returnData.push({
-								json: { error: (error as JsonObject).message },
+								json: { error: sanitizeErrorMessage(error) },
 								pairedItem: fallbackPairedItems ?? [{ item: i }],
 							});
 							continue;
@@ -795,7 +807,7 @@ export class MongoDb implements INodeType {
 					} catch (error) {
 						if (this.continueOnFail()) {
 							returnData.push({
-								json: { error: (error as JsonObject).message },
+								json: { error: sanitizeErrorMessage(error) },
 								pairedItem: fallbackPairedItems ?? [{ item: i }],
 							});
 							continue;
@@ -823,7 +835,7 @@ export class MongoDb implements INodeType {
 					} catch (error) {
 						if (this.continueOnFail()) {
 							returnData.push({
-								json: { error: (error as JsonObject).message },
+								json: { error: sanitizeErrorMessage(error) },
 								pairedItem: fallbackPairedItems ?? [{ item: i }],
 							});
 							continue;
@@ -832,6 +844,11 @@ export class MongoDb implements INodeType {
 					}
 				}
 			}
+		} catch (error) {
+			const sanitizedMessage = sanitizeErrorMessage(error);
+			if (error instanceof Error && sanitizedMessage === error.message) throw error;
+
+			throw new NodeOperationError(node, sanitizedMessage);
 		} finally {
 			await client.close().catch(() => {});
 		}
